move follow components into contacts.jsx
