+++ ui/views/contacts.jsx
@@ -1,13 +1,12 @@
 'use babel'
 import React from 'react'
 import pull from 'pull-stream'
+import ip from 'ip'
 import { Link } from 'react-router'
 import VerticalFilledContainer from 'patchkit-vertical-filled'
 import { UserSummaries } from '../com/user/summary'
 import LeftNav from '../com/leftnav'
 import RightNav from '../com/rightnav'
-import FollowNearby from '../com/forms/follow-nearby'
-import FollowFoafs from '../com/forms/follow-foafs'
 import PubInvite from 'patchkit-form-pub-invite'
 import ModalSingle from 'patchkit-modal/single'
 import social from '../lib/social-graph'
@@ -87,6 +86,53 @@ class Pubs extends React.Component {
   }
 }
 
+// peers on the wifi
+class FollowNearby extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state ={
+      foundNearbyPeers: [] // array of IDs
+    }
+  }
+  componentDidMount() {
+    // update peers after a delay
+    // (use a delay so the user sees the app "scanning")
+    this.onPeersUpdate = () => {
+      var peers = new Set()
+      app.peers
+        .filter(p => !ip.isLoopback(p.host) && ip.isPrivate(p.host))
+        .filter(p => !social.flags(app.user.id, p.key))
+        .forEach(p => peers.add(p.key))
+      peers = [...peers]
+      this.setState({ foundNearbyPeers: peers })
+
+      // HACK
+      // I want names to show up on these peers, and the local peers may still be replicating messages
+      // so I'm having the app refresh its state every time the peers have polled
+      // a better solution would be to have the backend state in sync with frontend through a continuous 'change' stream
+      // but that's a future project
+      // -prf
+      app.fetchLatestState()
+    }
+    this.onPeersUpdate()
+    app.on('update:peers', this.onPeersUpdate)
+  }
+
+  componentWillUnmount() {
+    app.removeListener('update:peers', this.onPeersUpdate)
+  }
+
+  render() {
+    const peers = this.state.foundNearbyPeers
+    const hasPeers = (peers.length > 0)
+    return <div>
+      <h1>Nearby</h1>
+      <h3 style={{marginTop: 5}}>{ hasPeers ? 'Potential contacts on your WiFi.' : 'Nobody found on your WiFi.' }</h3>
+      <UserSummaries ids={peers} />
+    </div>
+  }
+}
+
 class Friends extends React.Component {
   constructor(props) {
     super(props)
@@ -138,6 +184,45 @@ class Follows extends React.Component {
       <h1>Following</h1>
       <h3>Users you follow, but who {"don't"} follow you back. They may not see messages from you.</h3>
       <UserSummaries ids={this.state.follows} />
+    </div>
+  }
+}
+
+class FollowFoafs extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state ={
+      foafs: [] // array of IDs
+    }
+  }
+  componentDidMount() {    
+    // get foafs
+    var hits = {}
+    pull(
+      app.ssb.friends.createFriendStream({ hops: 2 }),
+      pull.filter(id => {
+        // remove already-followed, flagged, self, duplicates, and users w/o names and pics
+        if (hits[id]) return false
+        hits[id] = true
+        return id !== app.user.id && !!app.users.names[id] && !!u.profilePic(id) && !social.follows(app.user.id, id) && !social.flags(app.user.id, id)
+      }),
+      pull.collect((err, ids) => {
+        if (err)
+          return app.minorIssue('An error occurred while fetching users', err)
+        ids.sort((a, b) => social.followers(b).length - social.followers(a).length)
+        this.setState({ foafs: ids })
+      })
+    )
+  }
+
+  render() {
+    const foafs = this.state.foafs
+    if (foafs.length === 0)
+      return <div/>
+    return <div>
+      <h1>Friends of Friends</h1>
+      <h3 style={{marginTop: 5}}>Potential contacts from your social network.</h3>
+      <UserSummaries ids={foafs} />
     </div>
   }
 }
